Give the user a chance to back-out of the installation after displaying a list of files which will be created or modified to fix #6

+++ bin/lint.js
@@ -4,6 +4,7 @@ const fs = require('fs')
 const path = require('path')
 const mergeJson = require('../src/helpers/mergejson')
 const argv = require('minimist')(process.argv.slice(2))
+const inquirer = require('inquirer')
 
 const { COPYFILE_EXCL } = fs.constants
 
@@ -24,25 +25,61 @@ if (!stacks.length) {
     return
 }
 
-['all'].concat(stacks).forEach((stack) => {
+let stackFileList = []
+
+stacks.forEach((stack) => {
     const stackPathJson = path.resolve(__dirname, '../src/stacks/', stack, 'json')
     const stackPathStubs = path.resolve(__dirname, '../src/stacks/', stack, 'stubs')
 
     if (fs.existsSync(stackPathStubs)) {
-        const stubFiles = fs.readdirSync(stackPathStubs)
-
-        stubFiles.forEach((fileName) => {
-            if (fs.existsSync(fileName)) return
-
-            fs.copyFileSync(path.resolve(stackPathStubs, fileName), fileName, COPYFILE_EXCL)
-        })
+        stackFileList = stackFileList.concat(fs.readdirSync(stackPathStubs))
     }
 
     if (fs.existsSync(stackPathJson)) {
-        const stackFiles = fs.readdirSync(stackPathJson)
-
-        stackFiles.forEach((fileName) => {
-            mergeJson.files(path.resolve(stackPathJson, fileName), fileName)
-        })
+        stackFileList = stackFileList.concat(fs.readdirSync(stackPathJson))
     }
+})
+
+stackFileList = Array.from(new Set(stackFileList)).sort().join('\n')
+
+inquirer.prompt([{
+    type: 'confirm',
+    name: 'install',
+    message: `
+The following files will be created/updated.
+${stackFileList}
+Are you sure you want to continue?
+`,
+    default: true,
+}]).then((answer) => {
+    if (!answer.install) {
+        console.log('No stacks were installed')
+        return false
+    }
+
+    ['all'].concat(stacks).forEach((stack) => {
+        const stackPathJson = path.resolve(__dirname, '../src/stacks/', stack, 'json')
+        const stackPathStubs = path.resolve(__dirname, '../src/stacks/', stack, 'stubs')
+
+        if (fs.existsSync(stackPathStubs)) {
+            const stubFiles = fs.readdirSync(stackPathStubs)
+
+            stubFiles.forEach((fileName) => {
+                if (fs.existsSync(fileName)) return
+
+                fs.copyFileSync(path.resolve(stackPathStubs, fileName), fileName, COPYFILE_EXCL)
+            })
+        }
+
+        if (fs.existsSync(stackPathJson)) {
+            const stackFiles = fs.readdirSync(stackPathJson)
+
+            stackFiles.forEach((fileName) => {
+                mergeJson.files(path.resolve(stackPathJson, fileName), fileName)
+            })
+        }
+    })
+
+    console.log('Stack linting is now installed. Please run npm/composer install as required.')
+    return true
 })
